Main/PCIDeviceAttachmentImpl.cpp: Signed/unsigned conversion issues. bugref:9790
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85245 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-all/PCIDeviceAttachmentImpl.cpp
@@ -93,15 +93,17 @@ HRESULT PCIDeviceAttachment::initCopy(IMachine *aParent, PCIDeviceAttachment *aT
 HRESULT PCIDeviceAttachment::i_loadSettings(IMachine *aParent,
                                             const settings::HostPCIDeviceAttachment &hpda)
 {
-    return init(aParent, hpda.strDeviceName, hpda.uHostAddress, hpda.uGuestAddress, TRUE);
+    /** @todo r=bird: Inconsistent signed/unsigned crap. */
+    return init(aParent, hpda.strDeviceName, (LONG)hpda.uHostAddress, (LONG)hpda.uGuestAddress, TRUE);
 }
 
 
 HRESULT PCIDeviceAttachment::i_saveSettings(settings::HostPCIDeviceAttachment &data)
 {
     Assert(m);
-    data.uHostAddress = m->HostAddress;
-    data.uGuestAddress = m->GuestAddress;
+    /** @todo r=bird: Inconsistent signed/unsigned crap. */
+    data.uHostAddress  = (uint32_t)m->HostAddress;
+    data.uGuestAddress = (uint32_t)m->GuestAddress;
     data.strDeviceName = m->DevName;
 
     return S_OK;
